+++ lib/cosmos_epub.dart
@@ -13,6 +13,10 @@ import 'package:flutter/services.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get_storage/get_storage.dart';
 
+import 'package:http/http.dart' as http;
+
+///TODO: Optimize with isolates
+
 class CosmosEpub {
   static final GlobalKey<NavigatorState> navigatorKey =
       GlobalKey<NavigatorState>();
@@ -29,9 +33,59 @@ class CosmosEpub {
       String chapterListTitle = 'Table of Contents',
       bool shouldOpenDrawer = false,
       int starterChapter = -1}) async {
-
-    ///TODO: Optimize with isolates
     var bytes = File(localPath).readAsBytesSync();
+    EpubBook epubBook = await EpubReader.readBook(bytes.buffer.asUint8List());
+
+    if (!context.mounted) return;
+    _openBook(
+        context: context,
+        epubBook: epubBook,
+        bookId: bookId,
+        shouldOpenDrawer: shouldOpenDrawer,
+        starterChapter: starterChapter,
+        chapterListTitle: chapterListTitle,
+        onPageFlip: onPageFlip,
+        onLastPage: onLastPage,
+        accentColor: accentColor);
+  }
+
+  static Future<void> openFileBook(
+      {required Uint8List bytes,
+      required BuildContext context,
+      required String bookId,
+      Color accentColor = Colors.indigoAccent,
+      Function(int currentPage, int totalPages)? onPageFlip,
+      Function(int lastPageIndex)? onLastPage,
+      String chapterListTitle = 'Table of Contents',
+      bool shouldOpenDrawer = false,
+      int starterChapter = -1}) async {
+    EpubBook epubBook = await EpubReader.readBook(bytes.buffer.asUint8List());
+
+    if (!context.mounted) return;
+    _openBook(
+        context: context,
+        epubBook: epubBook,
+        bookId: bookId,
+        shouldOpenDrawer: shouldOpenDrawer,
+        starterChapter: starterChapter,
+        chapterListTitle: chapterListTitle,
+        onPageFlip: onPageFlip,
+        onLastPage: onLastPage,
+        accentColor: accentColor);
+  }
+
+  static Future<void> openURLBook(
+      {required String urlPath,
+      required BuildContext context,
+      Color accentColor = Colors.indigoAccent,
+      Function(int currentPage, int totalPages)? onPageFlip,
+      Function(int lastPageIndex)? onLastPage,
+      required String bookId,
+      String chapterListTitle = 'Table of Contents',
+      bool shouldOpenDrawer = false,
+      int starterChapter = -1}) async {
+    final result = await http.get(Uri.parse(urlPath));
+    final bytes = result.bodyBytes;
     EpubBook epubBook = await EpubReader.readBook(bytes.buffer.asUint8List());
 
     if (!context.mounted) return;
@@ -57,9 +111,6 @@ class CosmosEpub {
       String chapterListTitle = 'Table of Contents',
       bool shouldOpenDrawer = false,
       int starterChapter = -1}) async {
-
-    ///TODO: Optimize with isolates
-
     var bytes = await rootBundle.load(assetPath);
     EpubBook epubBook = await EpubReader.readBook(bytes.buffer.asUint8List());
 
@@ -100,10 +151,7 @@ class CosmosEpub {
           epubBook: epubBook,
           starterChapter: starterChapter >= 0
               ? starterChapter
-              : bookProgress
-                      .getBookProgress(bookId)
-                      .currentChapterIndex ??
-                  0,
+              : bookProgress.getBookProgress(bookId).currentChapterIndex ?? 0,
           shouldOpenDrawer: shouldOpenDrawer,
           bookId: bookId,
           accentColor: accentColor,