+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
 * This is initial release of cosmos_epub package
+
+## 0.0.2
+
+* Build error fix
+* Add `CosmosEpub.openURLBook` & `CosmosEpub.openFileBook` methods